Initial commit
Found an old project on one of my machines and decided to rewrite it and start using it again. Figured I’d share it for anyone looking for a super minimal text editor. I may update or expand it later, but it already fits my everyday needs, and since finding it, I’ve been preferring it over Neovim.